Add a simple helm test (#1449)
* Add a simple helm test

This could be useful for Terraform configs check that installation
was completed, for example after switching to a new cluster.

* Add docs, import cleanup.

Described `helm test` command. Updated `crd-client` docs and version in
yaml of example Job which creates a GameServer and also used by the test.
Bump version of Agones in go.mod.

Co-authored-by: Mark Mandel <markmandel@google.com>