Flaky: TestGameServerReserve (#1565)
* Flaky: TestGameServerReserve

Trying to test transitive state while it is actively changing is
fraught with peril. Tests flake, hair is pulled, and screams echo
throughout the lands.

This PR allows one to send a duration to the RESERVE command to
udp-simple, as well as update TestGameServerReserve to only confirm
concrete states, which therefore can't flake.

Closes #1543

* Updated certs for TestTlsWebhook

Included instructions on how to update.